+++ client/include/webcface/internal/data_store1.h
@@ -1,9 +1,8 @@
 #pragma once
 #include <mutex>
 #include <optional>
-#include <deque>
-#include "webcface/log.h"
 #include "webcface/encoding/encoding.h"
+#include "webcface/field.h"
 
 WEBCFACE_NS_BEGIN
 namespace internal {
@@ -64,7 +63,6 @@ class SyncDataStore1 {
 
 #if WEBCFACE_SYSTEM_DLLEXPORT
 extern template class SyncDataStore1<std::string>; // test用
-extern template class SyncDataStore1<std::shared_ptr<std::deque<LogLineData>>>;
 extern template class SyncDataStore1<std::chrono::system_clock::time_point>;
 #endif