+++ client/src/data_store2.cc
@@ -249,7 +249,6 @@ StrMap2<unsigned int> SyncDataStore2<T, ReqT>::transferReq() {
     // }
 }
 
-// ライブラリ外からは参照できないが、testのためにexportしている
 template class SyncDataStore2<std::string, int>; // test用
 template class SyncDataStore2<ValueData, int>;
 template class SyncDataStore2<TextData, int>;
@@ -259,5 +258,6 @@ template class SyncDataStore2<RobotModelData, int>;
 template class SyncDataStore2<Canvas3DData, int>;
 template class SyncDataStore2<Canvas2DData, int>;
 template class SyncDataStore2<ImageData, message::ImageReq>;
+template class SyncDataStore2<std::shared_ptr<LogData>, int>;
 } // namespace internal
 WEBCFACE_NS_END